requirements.txt
pandas
scikit-learn
umap-learn
hdbscan
catboost
joblib
google-generativeai
numpy
